+++ packages/plugin-gallery/web/src/gallery-component.jsx
@@ -203,7 +203,7 @@ class GalleryComponent extends PureComponent {
   renderMobileNativeLoader = ({ url }) =>
     url ? null : (
       <div className={styles.mobileNativeLoaderContainer}>
-        <Loader />
+        <Loader type={'mini'} disableProgress />
       </div>
     );