+++ packages/plugin-table/web/src/table-component.jsx
@@ -7,8 +7,9 @@ import TableViewer from './table-viewer';
 import styles from '../statics/styles/table-component.scss';
 import Table from './domain/table';
 import { getRange } from './domain/tableDataUtil';
-import { createEmptyCellEditor } from './tableUtil';
-import { AddNewSection, TableToolbar, Columns, Rows } from './components';
+import { createEmptyCellEditor, handleCellClipboardEvent } from './tableUtil';
+import { AddNewSection, Columns, Rows } from './components';
+import TableToolbar from './TableToolbar/TableToolbar';
 import { isPluginFocused, TOOLBARS } from 'wix-rich-content-editor-common';
 import { isEmpty, isNumber } from 'lodash';
 import classNames from 'classnames';
@@ -468,6 +469,7 @@ class TableComponent extends React.Component {
               this.table.getSelectedRows(range)
             }
             t={t}
+            handleCellClipboardEvent={handleCellClipboardEvent}
           />
           <div className={styles.dragPreview} ref={this.dragPreview} />
         </div>